feat: select each YT renderer type
Different pages on youtube use different elements to display videos
all of which have the pattern `yt-*-renderer`. The structure between
them is similar enough that the code to get the title and channel name
is the same. So we select all of the different types and merge them
before filtering.